--- src/components/FeedModal/FeedModal.stories.js
@@ -1,8 +1,0 @@
-import FeedModal from './FeedModal';
-
-export default {
-  title: 'components/FeedModal',
-  component: FeedModal,
-};
-
-export const Default = (args) => <FeedModal {...args}></FeedModal>;
